+++ lib/espnscrape/ArrayConversions.rb
@@ -1,0 +1,48 @@
+# Array Extensions for Type Conversion
+class Array
+  # Create Hash Array
+  # @param keys [Array] Symbols to be used as field names
+  # @return [[Hash]] Array<Hash>
+  # @example
+  #   teams   = es.teamList # Array of Team data
+  #   hash_a  = teams.to_hashes
+  #   => [{:t_abbr=>"BOS", :t_name=>"Boston Celtics", :division=>"Atlantic", :conference=>"Eastern"} ... ]
+  def to_hashes(keys = [])
+    return [] if empty?
+    two_d = first.is_a? Array # Check for 2D array
+    keys = checkKeys(keys, two_d ? first.size : size) # Determine keys
+    return [Hash[keys.map.with_index { |key, idx| [key, self[idx]] }]] unless two_d # 1D Array
+    map { |ary| Hash[keys.map.with_index { |key, idx| [key, ary[idx]] }] }          # 2D Array
+  end
+
+  # Create Struct Array
+  # @param keys [Array] Symbols to be used as field names
+  # @return [[Struct]] Array<Struct>
+  # @example
+  #   teams   = es.teamList # Array of Team data
+  #   structs = teams.to_structs
+  #   => [#<struct t_abbr="BOS", t_name="Boston Celtics", division="Atlantic", conference="Eastern"> ... ]
+  def to_structs(keys = [])
+    return [] if empty?
+    keys = checkKeys(keys, first.size)
+    to_hashes(keys).map { |hash| Struct.new(*hash.keys).new(*hash.values) }
+  end
+
+  # Determine default field names
+  # @param keys [[Symbol]] Field Names
+  # @param k_id [Int] Key Identifier
+  # @return [[Symbol]]
+  def checkKeys(keys, k_id)
+    return keys unless keys.empty?
+    [BOX_P, BOX_T, GAME_F, GAME_P, ROSTER, TEAM_L].each do |default|
+      return default if default.size.eql?(k_id)
+    end
+  end
+
+  # Replace old symbol with new symbol in Array
+  # @param old_sym [Symbol] Symbol to remove
+  # @param new_sym [Symbol] Symbol to add
+  def change_sym!(old_sym, new_sym)
+    map! { |x| x.eql?(old_sym) ? new_sym : x }
+  end
+end